Update names for functional_form move to common.
