+++ drake/common/CMakeLists.txt
@@ -12,6 +12,7 @@ configure_file(drake_path.cc.in ${drake_path_cc})
 set(sources
   drake_assert.cc
   ${drake_path_cc}
+  functional_form.cc
   nice_type_name.cc
   text_logging.cc
   )
@@ -24,6 +25,7 @@ set(installed_headers
   drake_deprecated.h
   drake_path.h
   eigen_types.h
+  functional_form.h
   nice_type_name.h
   sorted_vectors_have_intersection.h
   text_logging.h
