+++ drake/common/test/functional_form_test.cc
@@ -1,4 +1,4 @@
-#include "drake/core/functional_form.h"
+#include "drake/common/functional_form.h"
 
 #include <cmath>
 #include <sstream>
